+++ src/TabManager.cpp
@@ -4,14 +4,14 @@ TabManager::TabManager()
 {
     for (size_t i = 0; i < coins.size(); ++i)
     {
-        coins[i] = std::make_unique<Coin>(i*20,0);
+        coins[i] = std::make_unique<Coin>((int)i*20,0);
     }
     
     const int offsetHeartTab = 400;
     
     for (size_t i = 0; i < hearts.size(); ++i)
     {
-        hearts[i] = std::make_unique<Heart>(offsetHeartTab + i * Tile::WIDTH, 0);
+        hearts[i] = std::make_unique<Heart>(offsetHeartTab + (int)i * Tile::WIDTH, 0);
     }
 }
 
@@ -19,7 +19,7 @@ void TabManager::update()
 {
     for (size_t i = 0; i < coins.size(); ++i)
     {
-        if (i < Player::coins)
+        if ((int)i < Player::coins)
             coins[i]->show = true;
         else
             coins[i]->show = false;
@@ -27,7 +27,7 @@ void TabManager::update()
     
     for (size_t i = 0; i < hearts.size(); ++i)
     {
-        if (i < Player::hp)
+        if ((int)i < Player::hp)
             hearts[i]->isRed = true;
         else
             hearts[i]->isRed = false;
